+++ CHANGES.md
@@ -74,6 +74,9 @@ v0.14.0 (unreleased)
 v0.13.4 (unreleased)
 --------------------
 
+* Fix bug that occurred in certain cases when using the state attribute limit
+  helper with a state class that did not have a log attribute. [#1842]
+
 * Fix HDF5 reader for string columns. [#1840]
 
 * Fix visual bug in link editor in advanced mode when resizing window. [#1837]